Documentation per branch and new directives
Add a directory ext which contains directives for rst files. FOr now
their is only directives about ros. Those directives are copied to
src-root and add in the pythonpath. So the user only need to add
"ros-directives" in the conf.py extension part to use it within this
action.
Doc is also created per branch to avoid deleting doc of peers :)